pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "acceldocs-backend"
version = "0.1.0"
description = "AccelDocs backend — Google Docs → Zensical publishing pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "markdownify>=0.13.0",
    "gitpython>=3.1.40",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-slugify>=8.0.0",
    "bleach>=6.1.0",
    "python-multipart>=0.0.7",
    "jinja2>=3.1.0",
    "zensical>=0.1.0",
    "markdown>=3.6",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
postgres = [
    "psycopg2-binary>=2.9.9",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
